backend/requirements.txt
Django>=5.0 
djangorestframework>=3.15 
pydantic>=2.8